requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
clip==1.0
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
decord==0.6.0
diffusers==0.35.2
dill==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
-e git+https://github.com/agwmon/frame-guidance.git@da3c05ae362fd3119799d398fcd6b0956fe08a84#egg=image_gen_aux&subdirectory=pipelines/utils/image_gen_aux
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
parameterized==0.9.0
peft==0.18.0
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.45.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spandrel==0.4.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.2.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
zipp==3.23.0
